LeetCode Submission - Koko Eating Bananas

+++ Koko Eating Bananas/Koko Eating Bananas.cpp
@@ -1,6 +1,6 @@
 class Solution {
 public:
-    int bananasFinished(vector<int>& piles, int speed){
+    double bananasFinished(vector<int>& piles, int speed){
         double total=0;
         for(int pile: piles){
             total+=pile / speed;
